// semFront/src/app/models/user-interface.ts
export interface UserInterface{
    first_name?: string;
    last_name?: string;
    first_last_name?: string;
    second_last_name?: string;
    fecha_nacimiento?: string;
    sexo?: string;
    password?: string;
    password2?: string;
    email?: string;
    telefono?: string;
    type?: string;
    id_sem?: any;
    tipo?: string;
    cedula?: string;
    estado?: string;
    especialidad?: string;
    caducidad?: string;
    room?: string;
    auditor?: string;
}