// semFront/src/app/models/resultado-interface.ts
export interface ResultadoInterface{
    id?: string;
    fecha_realizacion?: string;
    fechaEstudio?: string;
    tipoEstudio?: string;
    notasEstudio?: string;
    file?: string;
    fileDicom?: string;
    Paciente?: any;
    creador?: string;
    editor?: string;
    creacion?: string;
    ultimaActualizacion?: string;
    results ?: any;
}