// semFront/src/app/models/consulta-interface.ts
export interface ConsultaInterface{
    id?: string;
    peso?: string;
    talla?: string;
    indiceMasaCorporal?: string;
    frecuenciaCardiaca?: string;
    frecuenciaRespiratoria?: string;
    tensionnArterialSistole?: string;
    tensionArterialDiastole?: string;
    temperatura?: string;
    sexo?: string;
    subjetivo?: string;
    objetivo?: string;
    analisis?: string;
    clinicos?: string;
    impresionDiagnostica?: string;
    manejoRecomendaciones?: string;
    plan?: string;
    Paciente?: string;
    Medico?: any;
    creacion?: string;
    ultimaActualizacion?: string;
}