// semFront/src/app/models/historiaClinica-interface.ts
export interface HistoriaClinicaInterface{
    id?: string;
    entidad_nacimiento?: string;
    religion?: string;
    ocupacion?: string;
    calle?: string;
    colonia?: string;
    num_interior?: string;
    num_exterior?: string;
    cp?: string;
    municipio?: string;
    localidad?: string;
    estado?: string;
    telefonoCasa?: string;
    telefonoOficina?: string;
    telefonoCelular?: string;
    email?: string;
    hijos?: any;
    numHijos?: string;
    estadoCivil?: string;
    maxGradoEstudios?: string;
    estadoSalud?: string;
    tipoInterrogatorio?: string;
    informador?: string;
    sexo?: string;
    peso?: any;
    talla?: any;
    temperatura?: string;
    indiceMasaCorporal?: any;
    frecuenciaCardiaca?: string;
    frecuenciaRespiratoria?: string;
    tensionArterialSistole?: string;
    tensionArterialDiastole?: string;
    sistemaNeurologico?: string;
    sistemaTegumentario?: string;
    sistemaMusculoEsqueletico?: string;
    sistemaRespiratorio?: string;
    sistemaDigestivo?: string;
    sistemaUrinario?: string;
    sistemaReproductor?: string;
    sistemasSentidos?: string;
    antecedentesFamiliares?: string;
    antecedentesPersonalesNoPatologicos?: string;
    antecedentesPersonalesPatologicos?: string;
    procedimientoActual?: string;
    auxiliaresDiagnostico?: string;
    Paciente?: string;
    creacion?: string;
    creador?: string;
    editor?: string;
    ultimaActualizacion?: string;
    results ?: any;
    count?: any;
}