// semFront/src/app/models/tablaPermisos-interface.ts
export interface TablaPermisosInterface{
    id?: string;
	token?: string;
    caducidad?: string;
    email?: string;
    tipo?: string;
    estatus?: string;
    Medico ?: any;
    Paciente ?: any;
    creador?: string;
    editor?: string;
    creacion?: string;
    ultimaActualizacion?: string;
    results ?: any;
    count ?: any;
    length ?: any;
    }