// semFront/src/app/models/medico-interface.ts
export interface MedicoInterface{
    id?: string;
    medicoSemin?: string;
    verificado?: string;
    caducidad?: string;
    nombre?: string;
    apellido_paterno?: string;
    apellido_materno?: string;
    imagen?: string; 
    cedula?: string;
    especialidad?: string;
    subespecialidad?: string;
    sexo?: string;
    ranking?: string;
    estrellas?: number;
    creador?: string;
    editor?: string;
    fecha_nacimiento?: string;
    estatusConexion?: string;
    telefono?: string;
    calle?: string;
    colonia?: string;
    num_interior?: string;
    num_exterior?: string;
    cp?: string;
    municipio?: string;
    localidad?: string;
    estado?: string;
    hora_apertura?: string;
    hora_cierre?: string;
    hora_apertura_sab?: string;
    hora_cierre_sab?: string;
    hora_apertura_dom?: string;
    hora_cierre_dom?: string;
    descripcion?: string;
    email?: string;
    creacion?: string;
    ultimaActualizacion?: string;
    results?: any;
    count?: any;
}