// semFront/src/app/models/empleado-interface.ts
export interface EmpleadoInterface{
    id?: string;
    nombre?: string;
    apellido_paterno?: string;
    apellido_materno?: string;
    auditor?: string;
    directivo?: string;
    puesto?: string;
    sexo?: string;
    fecha_nacimiento?: string;
    telefono?: string;
    email?: string;
}